ds/async_await_test.go
package ds_test

import (
	"go-data-structure/ds"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestAsyncAwait(t *testing.T) {
	tests := []struct {
		name        string
		fn          func() (string, error)
		expectedRes string
		expectedErr error
	}{
		{
			name: "test async-await",
			fn: func() (string, error) {
				req := ds.Async(func() (string, error) {
					return "completed", nil
				})

				return req.Await()
			},
			expectedRes: "completed",
		},
	}

	for _, test := range tests {
		res, err := test.fn()
		assert.Equal(t, test.expectedErr, err)
		assert.Equal(t, test.expectedRes, res)
	}
}
